Add row identifier
We're attempting to provide an identifier so the user can fix the
data back in their system at the source. Becky suggested these
combination of fields to get the TAS.

Currently, the logic is if the field doesn't exist in the data, just
not include it. This might not be the best logic because it
make make the number more confusing. Another option is to include
the number 0 where the data doesn't exist. This depends on what
the user would expect.