Separated out identifier and try at design
The identifier for a row was changed so TAS was a single entity with
"-" between each part. For every file that has them, separate
identifiers were added.

I attempted to create a better hierarchy and grouping in the design by
playing around with space, color and separating elements like borders.
I'm hoping to have this reviewed by a design professional.